requirements/common.txt
django~=3.0.0
djangorestframework
markdown
django-filter
pyjwt